Release v0.1.1 (#2)
## Summary

- add encrypted custom FoundationDB catalog support and coverage
- stabilize the catalog suite and resolve Clippy/style debt
- upgrade DuckDB and DuckLake upstream pins
- fix three-platform release packaging and portable checksums

## Validation

- `just ducklake-fdb-release-gate`
- `just style-check`
- `just ducklake-release`
- checksums and binary architectures verified for Linux amd64, Linux
arm64, and macOS arm64

+++ crates/ducklake-catalog/tests/upstream_partitioning_mirrors.rs
@@ -96,7 +96,7 @@ fn mirrors_bucket_pruning_test_partition_scan_uses_bucket_partition_values() {
         &mut kv,
         catalog,
         (0..10)
-            .map(|bucket| data_file(bucket + 1, table, bucket as u64 * 100, 100))
+            .map(|bucket| data_file(bucket + 1, table, bucket * 100, 100))
             .collect(),
         Vec::new(),
         &[],
